RAG Agent/requirements.txt
openai>=1.40.0
pinecone>=7.0.0
python-dotenv>=1.0.1
pypdf>=4.3.1
langchain-text-splitters>=0.0.2
tiktoken>=0.7.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
tqdm>=4.66.0

# Auth / DB
SQLAlchemy>=2.0.30
bcrypt>=4.1.3
PyJWT>=2.8.0
